+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add pager support on Windows
 - Add Intel RSD Network card information decoder #81
 - Add Intel RSD PCIe information decoder #82
+- Add Intel RSD TPM information decoder #85
 
 ### Fixed